+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - **NEW FEATURE**: Broadcast channel can be displayed with `showBroadcast`
 - Reduce unnecessary API calls
+- Minor revisions to README
 - Small cosmetic tweaks
 
 ## [2.1.1](https://github.com/dathbe/MMM-PGA/compare/v2.1.0...v2.1.1) - 2025-04-28